Make intellijPublishUsername and intellijPublishPassword optional
The current guide with directly referencing `intellijPublishUsername` and `intellijPublishPassword` is somehow unpractical for plugin contributors that do not want to publish the plugin but expect the plugin project to work out of the box after checkout.  When they try to open the gradle project in IDEA they are faced with error messages like:
```
[...]
> Could not get unknown property 'intellijPublishUsername' for task ':publishPlugin' of type org.jetbrains.intellij.tasks.PublishTask.
[...]
```
Users without gradle experience might give up at this point.  Therefore it's better to make these parameters optional.