Document incompatibilities
During the decoupling of ide.core and ide.core.impl modules from platform.ide and platofrm.ide.impl there were some incompatible changes that couldn't be avoided.
The main reason for it was that there were dependency on the UI sepcific implementations in the API, which in general should be avoided.