kotlin.md: remove "Kotlin Code FAQ" section
It contains a single question, and it doesn't look important. If it is needed, it should be moved to the Plugin Development FAQ (which is planned to be removed), where we have questions about Java and XML, so Kotlin as another language fits there. I woudl stick to removal, though.