Update welcome.md
Earlier we only had inline images in links, so it was possible to assign proper styles to images that are used in this way; now we've implemented banners that are block images that can also be found in links. Now the image type must be explicitly specified, which is what I did here.